Scissoring the input of the scrolllist works perfectly now!
git-svn-id: https://paroxysm.googlecode.com/svn/trunk@159 1a4693bc-ef48-11de-b206-85b67c77eac7

+++ source/ScrollList.h
@@ -33,7 +33,7 @@ class ScrollList : public Widget
 
         void setFontSize(int inSize);
 
-        void onMouseChange(float inX, float inY);
+        void onMouseChange(int inX, int inY);
         virtual void display();
 
         void findPixels(const Point2D<int>& inDisplay, float inRange);
